+++ autos/services/price_parsing/infrastructure/add_spare_task.py
