+++ autos/tasks.py
@@ -1,14 +1,9 @@
-from celery import shared_task
-from celery_singleton import Singleton
-from django.contrib.auth.models import User
-
 from autos.services.plotting.infrastructure.making_graphs_task import make_graphs_for_all_spares, make_graph_for_spare
 from autos.services.price_parsing.infrastructure.add_spare_task import add_spare
 from autos.services.price_parsing.infrastructure.parsing_task import make_request
 from celery_app import app
 
 
-# @shared_task(base=Singleton)
 @app.task(bind=True, max_retries=3)
 def do_make_request(self, user: str):
     """Выполнения новых запросов для всех запчастей"""
