+++ docker/Dockerfile
@@ -1,0 +1,24 @@
+FROM python:3.12
+
+WORKDIR /autoworld
+EXPOSE 8000
+
+ENV PYTHONDONTWRITEBYTECODE=1 PYTHONUNBUFFERED=1
+
+RUN apt-get update && \
+    apt-get install -y locales && \
+    sed -i -e 's/# ru_RU.UTF-8 UTF-8/ru_RU.UTF-8 UTF-8/' /etc/locale.gen && \
+    dpkg-reconfigure --frontend=noninteractive locales
+
+ENV LANG ru_RU.UTF-8
+ENV LC_ALL ru_RU.UTF-8
+
+#RUN apt-get install -y wget
+#RUN wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add - \
+#    && echo "deb http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google.list
+#RUN apt-get update && apt-get -y install google-chrome-stable
+
+COPY . /autoworld
+
+COPY ./requirements.txt /temp/requirements.txt
+RUN pip install --no-cache-dir -r /temp/requirements.txt
